pyproject.toml
[project]
name = "kb-mcp"
version = "0.1.0"
description = "AI cross-context synchronization layer — shared knowledge base MCP server for Claude Code, Copilot, and Codex"
license = "MIT"
readme = "README.md"
requires-python = ">=3.12"
authors = [
    { name = "okash1n" },
]
keywords = ["mcp", "knowledge-base", "ai", "claude", "copilot", "codex", "obsidian"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Documentation",
]
dependencies = [
    "mcp[cli]>=1.26.0",
    "python-ulid>=3.0.0",
    "pyyaml>=6.0",
]

[project.urls]
Homepage = "https://github.com/okash1n/kb"
Repository = "https://github.com/okash1n/kb"

[project.scripts]
kb-mcp = "kb_mcp.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/kb_mcp"]
